backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
pydantic-settings
sqlalchemy
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
stripe
httpx
email-validator
python-dateutil
